+++ user_task_list.go
@@ -2,7 +2,8 @@ package asana
 
 import "fmt"
 
-// UserTaskList represents the tasks assigned to a particular user.
+// UserTaskList represents the tasks assigned to a particular user. It provides
+// API access to a user's My Tasks view in Asana.
 type UserTaskList struct {
 	// Read-only. Globally unique ID of the object
 	ID string `json:"gid,omitempty"`
@@ -10,14 +11,12 @@ type UserTaskList struct {
 	// Read-only. The name of the object.
 	Name string `json:"name,omitempty"`
 
-	// The owner of the user task list.
+	// Read-only. The owner of the user task list, i.e. the person whose My
+	// Tasks is represented by this resource.
 	Owner *User `json:"owner,omitempty"`
 
-	// Read-only. Workspaces and organizations this user may access.
-	//
-	// Note: The API will only return workspaces and organizations that also
-	// contain the authenticated user.
-	Workspaces []*Workspace `json:"workspaces,omitempty"`
+	// Read-only. The workspace in which the user task list is located.
+	Workspace *Workspace `json:"workspace,omitempty"`
 }
 
 // Fetch loads the full details for this UserTaskList
@@ -26,4 +25,27 @@ func (u *UserTaskList) Fetch(client *Client, options ...*Options) error {
 
 	_, err := client.get(fmt.Sprintf("/user_task_lists/%s", u.ID), nil, u, options...)
 	return err
+}
+
+// TaskList returns the user task list (My Tasks) for this user in the given
+// workspace
+func (u *User) TaskList(client *Client, workspace *Workspace, options ...*Options) (*UserTaskList, error) {
+	client.trace("Loading task list for user %q in workspace %q", u.ID, workspace.ID)
+
+	result := &UserTaskList{}
+
+	queryOptions := append([]*Options{{Workspace: workspace.ID}}, options...)
+	_, err := client.get(fmt.Sprintf("/users/%s/user_task_list", u.ID), nil, result, queryOptions...)
+	return result, err
+}
+
+// Tasks returns the compact list of tasks in this user task list
+func (u *UserTaskList) Tasks(client *Client, options ...*Options) ([]*Task, *NextPage, error) {
+	client.trace("Listing tasks in user task list %q", u.ID)
+
+	var result []*Task
+
+	// Make the request
+	nextPage, err := client.get(fmt.Sprintf("/user_task_lists/%s/tasks", u.ID), nil, &result, options...)
+	return result, nextPage, err
 }